+++ pkg/internal/common/fileio.go
