+++ pkg/internal/common/print.go
